Cache teams (#77)
This was super interesting. Because of duplicate teams, we kind of need
to know teams anywhere we use a teamId (so we can map it to the latest).
Basically, we call getLatestTeams almost everywhere. Even a simple call
turns into maybe 30 requests for this one function. I was surprised by
this, but the more I thought about it, this was a mistake to do it this
way. A lot of these functions should have taken a TeamList and not make
the call themselves. Very dangerous indeed

who has the time to fix all that tho, so lets cache the teams for now.
Duplicate calls will now be served through the cache and our reads will
go down dramatically. Not happy, but ill take the optimization either
way. This was going to be called a lot regardless tbh.